+++ nbtester/__init__.py
@@ -1,1 +1,1 @@
-from .loader import load_cells, matplotlib_test  # NOQA
+from .loader import load_cells  # NOQA
